% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.sic_bootstrap}
\alias{print.sic_bootstrap}
\title{Print Bootstrap Standard Errors}
\usage{
\method{print}{sic_bootstrap}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An object of class \code{"sic_bootstrap"}.}

\item{digits}{Integer. Number of digits. Default: \code{4}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Bootstrap Standard Errors
}
